add class

+++ app/src/main/java/com/example/androidproject/Business.java
@@ -33,7 +33,7 @@ public class Business {
     private List<Category> categories = null;
     @SerializedName("rating")
     @Expose
-    private Integer rating;
+    private Double rating;
     @SerializedName("coordinates")
     @Expose
     private Coordinates coordinates;
@@ -82,7 +82,7 @@ public class Business {
      * @param isClosed
      * @param coordinates
      */
-    public Business(String id, String alias, String name, String imageUrl, Boolean isClosed, String url, Integer reviewCount, List<Category> categories, Integer rating, Coordinates coordinates, List<Object> transactions, String price, Location location, String phone, String displayPhone, Double distance) {
+    public Business(String id, String alias, String name, String imageUrl, Boolean isClosed, String url, Integer reviewCount, List<Category> categories, Double rating, Coordinates coordinates, List<Object> transactions, String price, Location location, String phone, String displayPhone, Double distance) {
         super();
         this.id = id;
         this.alias = alias;
@@ -166,11 +166,11 @@ public class Business {
         this.categories = categories;
     }
 
-    public Integer getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
